Add data-to-graph conversion functions for library
This change adds basic versions of functions used to convert data received into
chartJS compatible data and displaying it. This has been structured into
different files with some test data for now. External libraries in the test
folder are included via CDN for now but will changed in the future.