Convert to proper npm module
This change converts the `index.js` and `sessionsConverter.js` file to
proper npm modules which can be called by `require`. This also changes a
few method names as well as adding a new library (`chartjs-color`). It
also saves the exact version numbers so to protect against possible
updates which don't follow SemVer principles