Add linting
This change adds the `eslint` tool to the project as well as setting up
the configuration to be airbnb es5 rules. Additionally, it has added a
script to run this in `package.json`, running it over the `src` folder.
Lastly, this change also fixes the errors and warnings produced by the
linting tool, ensuring clean code. The bundles have then been updated to
show the updated changes.